+++ front/app/modules/commercial/custom_style/admin/components/CustomStyleInputs.tsx
@@ -2,7 +2,6 @@ import React from 'react';
 
 import RangeInput from 'components/UI/RangeInput';
 import messages from 'containers/Admin/settings/messages';
-import { ColorPickerSectionField } from 'containers/Admin/settings/customize';
 import { IAppConfigurationStyle } from 'services/appConfiguration';
 import { Label, ColorPickerInput } from 'cl2-component-library';
 import { FormattedMessage } from 'utils/cl-intl';
@@ -22,7 +21,7 @@ const CustomStyleInputs = ({
 }: Props) => {
   return (
     <>
-      <ColorPickerSectionField>
+      <SectionField>
         <Label>
           <FormattedMessage {...messages.imageOverlayColor} />
         </Label>
@@ -34,7 +33,7 @@ const CustomStyleInputs = ({
           }
           onChange={onChange('signedOutHeaderOverlayColor')}
         />
-      </ColorPickerSectionField>
+      </SectionField>
       <SectionField>
         <Label>
           <FormattedMessage {...messages.imageOverlayOpacity} />
